Add files via upload
This is my first introduction to python 3. In this notebook which is a part of the course python for Data Science, AI, and Development. I wrote my first line of code which was the "Hello, World!" code. The notebook delved into types of data in python, how to convert between said data types, and expressions as well as variables. This is the first notebook and I hope to publish many more in this learning journey I have undertaken.